Rename and move batchRegistration package into samples.registration.batch
