+++ vaadinfrontend/src/main/java/life/qbic/datamanager/views/projects/project/samples/registration/batch/MetaDataTypes.java
@@ -1,4 +1,4 @@
-package life.qbic.datamanager.views.projects.project.samples.batchRegistration;
+package life.qbic.datamanager.views.projects.project.samples.registration.batch;
 
 
 /**
